added format, theme for code and fixed issues
