+++ core/agent_harness/session/memory_extraction.py
@@ -3,8 +3,8 @@
 One best-effort LLM pass over a chat transcript. Callers schedule it via
 :func:`schedule_memory_extraction` after every recorded turn and again on
 session close / rotation. Mid-session runs coalesce onto a single daemon worker
-so rapid turns do not pile up provider calls. Process-exit close runs extraction
-synchronously (after resources are released) so durable facts always persist.
+so rapid turns do not pile up provider calls. Process-exit close waits for
+extraction to finish (interruptible by Ctrl+C) so durable facts always persist.
 Never raises out: any failure (LLM unavailable, malformed output, disk errors)
 is logged and ignored. Environment gates can disable the whole feature or only
 the extraction pass.
@@ -34,10 +34,10 @@ logger = logging.getLogger(__name__)
 
 MIN_CHAT_MESSAGES = 2
 MAX_MEMORIES_PER_SESSION = 5
-# Upper bound on how long session-close extraction may delay shutdown. The
-# blocking LLM call runs off the main thread and is abandoned past this, so a
-# slow provider cannot hang exit or force the user to Ctrl+C out.
-_CLOSE_EXTRACTION_TIMEOUT_SECONDS = 5.0
+# Poll interval while waiting for close-path extraction. Short enough that
+# Ctrl+C stays responsive; the join itself runs until the worker finishes so
+# durable facts are never abandoned on a slow provider.
+_CLOSE_EXTRACTION_POLL_SECONDS = 0.25
 _MAX_TRANSCRIPT_TURNS = 30
 
 _FENCED_JSON_RE = re.compile(r"```(?:json)?\s*(\[.*?\])\s*```", re.DOTALL)
@@ -163,10 +163,10 @@ def schedule_memory_extraction(
         with _worker_lock:
             _pending_messages = None
             _pending_context = None
-        # Run the blocking extraction off the main thread and wait a bounded time
-        # for it. Shutdown never hangs on a slow LLM call, and the interruptible
-        # join keeps a Ctrl+C during exit from raising a stray SystemExit through
-        # the network read. The daemon thread is abandoned if it overruns.
+        # Run extraction off the main thread and wait until it finishes so
+        # durable facts always land before process exit. Poll the join so
+        # Ctrl+C during shutdown stays interruptible without raising through
+        # the network read mid-call.
         ctx = contextvars.copy_context()
         worker = threading.Thread(
             target=ctx.run,
@@ -175,7 +175,15 @@ def schedule_memory_extraction(
             daemon=True,
         )
         worker.start()
-        worker.join(timeout=_CLOSE_EXTRACTION_TIMEOUT_SECONDS)
+        try:
+            while worker.is_alive():
+                worker.join(timeout=_CLOSE_EXTRACTION_POLL_SECONDS)
+        except KeyboardInterrupt:
+            logger.warning(
+                "Memory extraction interrupted during session close; "
+                "final transcript facts may be incomplete"
+            )
+            raise
         return
     _schedule_coalesced(snapshot)
